menu as a separate component
